feat: split course.tsx's functionality into multiple files
