+++ hwproj.front/src/components/Courses/Course.tsx
@@ -17,9 +17,7 @@ import {
     Menu,
     MenuItem,
     Stack,
-    Typography,
-    TextField,
-    DialogActions
+    Typography
 } from "@mui/material";
 import {CourseExperimental} from "./CourseExperimental";
 import {useParams, useNavigate} from 'react-router-dom';
@@ -31,13 +29,11 @@ import {QRCodeSVG} from 'qrcode.react';
 import ErrorsHandler from "components/Utils/ErrorsHandler";
 import {useSnackbar} from 'notistack';
 import QrCode2Icon from '@mui/icons-material/QrCode2';
+import MailOutlineIcon from '@mui/icons-material/MailOutline';
 import {MoreVert} from "@mui/icons-material";
 import {DotLottieReact} from "@lottiefiles/dotlottie-react";
-import Avatar from "@material-ui/core/Avatar";
-import MailOutlineIcon from '@mui/icons-material/MailOutline';
-import {makeStyles} from '@material-ui/core/styles';
-import Autocomplete from '@mui/material/Autocomplete';
-import ValidationUtils from "../Utils/ValidationUtils";
+import InviteStudentDialog from "./InviteStudentDialog";
+import {makeStyles} from "@material-ui/core/styles";
 
 type TabValue = "homeworks" | "stats" | "applications"
 
@@ -60,415 +56,6 @@ interface IPageState {
     tabValue: TabValue
 }
 
-interface InviteStudentDialogProps {
-    courseId: number;
-    open: boolean;
-    onClose: () => void;
-    onStudentInvited: () => Promise<void>;
-    email?: string;
-}
-
-interface RegisterStudentDialogProps {
-    courseId: number;
-    open: boolean;
-    onClose: () => void;
-    onStudentRegistered: () => Promise<void>;
-    initialEmail?: string;
-}
-
-const RegisterStudentDialog: FC<RegisterStudentDialogProps> = ({courseId, open, onClose, onStudentRegistered, initialEmail}) => {
-    const classes = useStyles();
-    const {enqueueSnackbar} = useSnackbar();
-    const [email, setEmail] = useState(initialEmail || "");
-    const [name, setName] = useState("");
-    const [surname, setSurname] = useState("");
-    const [middleName, setMiddleName] = useState("");
-    const [errors, setErrors] = useState<string[]>([]);
-    const [isRegistering, setIsRegistering] = useState(false);
-
-    useEffect(() => {
-        if (initialEmail) {
-            setEmail(initialEmail);
-        }
-    }, [initialEmail]);
-
-    const registerStudent = async () => {
-        setIsRegistering(true);
-        setErrors([]);
-        try {
-            await ApiSingleton.coursesApi.coursesInviteStudent({
-                courseId: courseId,
-                email: email,
-                name: name,
-                surname: surname,
-                middleName: middleName
-            });
-            enqueueSnackbar("Студент успешно зарегистрирован и приглашен", {variant: "success"});
-            onClose();
-            await onStudentRegistered();
-        } catch (error) {
-            const responseErrors = await ErrorsHandler.getErrorMessages(error as Response);
-            if (responseErrors.length > 0) {
-                setErrors(responseErrors);
-            } else {
-                setErrors(['Не удалось зарегистрировать студента']);
-            }
-        } finally {
-            setIsRegistering(false);
-        }
-    };
-
-    return (
-        <Dialog
-            open={open}
-            onClose={() => !isRegistering && onClose()}
-            maxWidth="sm"
-            fullWidth
-        >
-            <DialogTitle>
-                <Grid container>
-                    <Grid item container direction={"row"} justifyContent={"center"}>
-                        <Avatar className={classes.avatar} style={{color: 'white', backgroundColor: '#00AB00'}}>
-                            <MailOutlineIcon/>
-                        </Avatar>
-                    </Grid>
-                    <Grid item container direction={"row"} justifyContent={"center"}>
-                        <Typography variant="h5">
-                            Зарегистрировать студента
-                        </Typography>
-                    </Grid>
-                </Grid>
-            </DialogTitle>
-            <DialogContent>
-                {errors.length > 0 && (
-                    <Typography color="error" align="center" style={{marginBottom: '16px'}}>
-                        {errors[0]}
-                    </Typography>
-                )}
-                <form className={classes.form}>
-                    <Grid container spacing={2}>
-                        <Grid item xs={12}>
-                            <TextField
-                                required
-                                fullWidth
-                                type="email"
-                                label="Электронная почта"
-                                variant="outlined"
-                                size="small"
-                                value={email}
-                                onChange={(e) => setEmail(e.target.value)}
-                                InputProps={{
-                                    autoComplete: "new-email"
-                                }}
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <TextField
-                                required
-                                fullWidth
-                                label="Имя"
-                                variant="outlined"
-                                size="small"
-                                value={name}
-                                onChange={(e) => setName(e.target.value)}
-                                InputProps={{
-                                    autoComplete: "new-name"
-                                }}
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <TextField
-                                required
-                                fullWidth
-                                label="Фамилия"
-                                variant="outlined"
-                                size="small"
-                                value={surname}
-                                onChange={(e) => setSurname(e.target.value)}
-                                InputProps={{
-                                    autoComplete: "new-surname"
-                                }}
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <TextField
-                                fullWidth
-                                label="Отчество"
-                                variant="outlined"
-                                size="small"
-                                value={middleName}
-                                onChange={(e) => setMiddleName(e.target.value)}
-                                InputProps={{
-                                    autoComplete: "new-middlename"
-                                }}
-                            />
-                        </Grid>
-                    </Grid>
-                    <Grid
-                        direction="row"
-                        justifyContent="flex-end"
-                        alignItems="flex-end"
-                        container
-                        style={{marginTop: '16px'}}
-                    >
-                        <Grid item>
-                            <Button
-                                variant="text"
-                                onClick={registerStudent}
-                                disabled={!email || !name || !surname || isRegistering}
-                                style={{
-                                    borderRadius: '8px',
-                                    textTransform: 'none',
-                                    fontWeight: 'normal',
-                                    color: '#3f51b5',
-                                    transition: 'opacity 0.3s ease',
-                                    opacity: (!email || !name || !surname || isRegistering) ? 0.5 : 1,
-                                    padding: '6px 16px',
-                                    marginRight: '8px'
-                                }}
-                            >
-                                {isRegistering ? 'Регистрация...' : 'Зарегистрировать'}
-                            </Button>
-                        </Grid>
-                        <Grid item>
-                            <Button
-                                variant="text"
-                                onClick={onClose}
-                                disabled={isRegistering}
-                                style={{
-                                    borderRadius: '8px',
-                                    textTransform: 'none',
-                                    fontWeight: 'normal',
-                                    color: '#3f51b5',
-                                    transition: 'opacity 0.3s ease',
-                                    opacity: isRegistering ? 0.5 : 1,
-                                    padding: '6px 16px'
-                                }}
-                            >
-                                Закрыть
-                            </Button>
-                        </Grid>
-                    </Grid>
-                </form>
-            </DialogContent>
-        </Dialog>
-    );
-};
-
-const InviteStudentDialog: FC<InviteStudentDialogProps> = ({courseId, open, onClose, onStudentInvited, email: initialEmail}) => {
-    const classes = useStyles();
-    const {enqueueSnackbar} = useSnackbar();
-    const [email, setEmail] = useState(initialEmail || "");
-    const [errors, setErrors] = useState<string[]>([]);
-    const [isInviting, setIsInviting] = useState(false);
-    const [showRegisterDialog, setShowRegisterDialog] = useState(false);
-    const [students, setStudents] = useState<AccountDataDto[]>([]);
-
-    const getStudents = async () => {
-        try {
-            const data = await ApiSingleton.accountApi.accountGetAllStudents();
-            setStudents(data);
-        } catch (error) {
-            console.error("Error fetching students:", error);
-        }
-    };
-
-    useEffect(() => {
-        getStudents();
-    }, []);
-
-    const getCleanEmail = (input: string) => {
-        return input.split(' / ')[0].trim();
-    };
-
-    const inviteStudent = async () => {
-        setIsInviting(true);
-        setErrors([]);
-        try {
-            const cleanEmail = getCleanEmail(email);
-            await ApiSingleton.coursesApi.coursesInviteStudent({
-                courseId: courseId,
-                email: cleanEmail,
-                name: "",
-                surname: "",
-                middleName: ""
-            });
-            enqueueSnackbar("Студент успешно приглашен", {variant: "success"});
-            setEmail("");
-            onClose();
-            await onStudentInvited();
-        } catch (error) {
-            const responseErrors = await ErrorsHandler.getErrorMessages(error as Response);
-            if (responseErrors.length > 0) {
-                setErrors(responseErrors);
-            } else {
-                setErrors(['Студент с такой почтой не найден']);
-            }
-        } finally {
-            setIsInviting(false);
-        }
-    };
-
-    const hasMatchingStudent = () => {
-        const cleanEmail = getCleanEmail(email);
-        return students.some(student => 
-            student.email === cleanEmail ||
-            `${student.surname} ${student.name}`.includes(cleanEmail)
-        );
-    };
-    
-    return (
-        <>
-            <Dialog
-                open={open}
-                onClose={() => !isInviting && onClose()}
-                maxWidth="sm"
-                fullWidth
-            >
-                <DialogTitle>
-                    <Grid container>
-                        <Grid item container direction={"row"} justifyContent={"center"}>
-                            <Avatar className={classes.avatar} style={{color: 'white', backgroundColor: '#00AB00'}}>
-                                <MailOutlineIcon/>
-                            </Avatar>
-                        </Grid>
-                        <Grid item container direction={"row"} justifyContent={"center"}>
-                            <Typography variant="h5">
-                                Пригласить студента
-                            </Typography>
-                        </Grid>
-                    </Grid>
-                </DialogTitle>
-                <DialogContent>
-                    {errors.length > 0 && (
-                        <Typography color="error" align="center" style={{marginBottom: '16px'}}>
-                            {errors[0]}
-                        </Typography>
-                    )}
-                    <form className={classes.form}>
-                        <Grid container spacing={2}>
-                            <Grid item xs={12}>
-                            <Autocomplete
-                                freeSolo
-                                options={students}
-                                getOptionLabel={(option) => 
-                                    typeof option === 'string' 
-                                        ? option 
-                                        : `${option.email} / ${option.surname} ${option.name}`
-                                }
-                                inputValue={email}
-                                onInputChange={(event, newInputValue) => {
-                                    setEmail(newInputValue);
-                                }}
-                                renderOption={(props, option) => (
-                                    <li {...props}>
-                                        <Grid container alignItems="center">
-                                            <Grid item>
-                                                <Box fontWeight="fontWeightMedium">
-                                                    {option.email} /
-                                                </Box>
-                                            </Grid>
-                                            <Grid item>
-                                                <Typography style={{marginLeft: '3px'}}>
-                                                    {option.surname} {option.name}
-                                                </Typography>
-                                            </Grid>
-                                        </Grid>
-                                    </li>
-                                )}
-                                renderInput={(params) => (
-                                    <TextField
-                                        {...params}
-                                        label="Введите email или ФИО"
-                                        variant="outlined"
-                                        size="small"
-                                        fullWidth
-                                    />
-                                )}
-                            />
-                            </Grid>
-                        </Grid>
-                        <Grid
-                            container
-                            spacing={1}
-                            justifyContent="flex-end"
-                            style={{marginTop: '16px'}}
-                        >
-                            <Grid item>
-                                <Button
-                                    variant="text"
-                                    onClick={() => {
-                                        setShowRegisterDialog(true);
-                                        onClose();
-                                    }}
-                                    disabled={hasMatchingStudent() || !getCleanEmail(email)}
-                                    style={{
-                                        borderRadius: '8px',
-                                        textTransform: 'none',
-                                        fontWeight: 'normal',
-                                        color: '#3f51b5',
-                                        transition: 'opacity 0.3s ease',
-                                        opacity: hasMatchingStudent() || !getCleanEmail(email) ? 0.5 : 1,
-                                        padding: '6px 16px',
-                                        marginRight: '8px'
-                                    }}
-                                >
-                                    Зарегистрировать
-                                </Button>
-                            </Grid>
-                            <Grid item>
-                                <Button
-                                    variant="text"
-                                    onClick={inviteStudent}
-                                    disabled={!hasMatchingStudent() || isInviting}
-                                    style={{
-                                        borderRadius: '8px',
-                                        textTransform: 'none',
-                                        fontWeight: 'normal',
-                                        color: '#3f51b5',
-                                        transition: 'opacity 0.3s ease',
-                                        opacity: !hasMatchingStudent() || isInviting ? 0.5 : 1,
-                                        padding: '6px 16px',
-                                        marginRight: '8px'
-                                    }}
-                                >
-                                    {isInviting ? 'Отправка...' : 'Пригласить'}
-                                </Button>
-                            </Grid>
-                            <Grid item>
-                                <Button
-                                    variant="text"
-                                    onClick={onClose}
-                                    disabled={isInviting}
-                                    style={{
-                                        borderRadius: '8px',
-                                        textTransform: 'none',
-                                        fontWeight: 'normal',
-                                        color: '#3f51b5',
-                                        transition: 'opacity 0.3s ease',
-                                        opacity: isInviting ? 0.5 : 1,
-                                        padding: '6px 16px'
-                                    }}
-                                >
-                                    Закрыть
-                                </Button>
-                            </Grid>
-                        </Grid>
-                    </form>
-                </DialogContent>
-            </Dialog>
-            
-            <RegisterStudentDialog
-                courseId={courseId}
-                open={showRegisterDialog}
-                onClose={() => setShowRegisterDialog(false)}
-                onStudentRegistered={onStudentInvited}
-                initialEmail={getCleanEmail(email)}
-            />
-        </>
-    );
-};
-
 const useStyles = makeStyles((theme) => ({
     paper: {
         marginTop: theme.spacing(3),
@@ -486,7 +73,7 @@ const useStyles = makeStyles((theme) => ({
     button: {
         marginTop: theme.spacing(1)
     },
-}))
+}));
 
 const Course: React.FC = () => {
     const {courseId, tab} = useParams()
@@ -576,6 +163,8 @@ const Course: React.FC = () => {
     }
 
     const getCourseFilesInfo = async () => {
+        // В случае, если сервис файлов недоступен, показываем пользователю сообщение
+        // и не блокируем остальную функциональность системы
         let courseFilesInfo = [] as FileInfoDTO[]
         try {
             courseFilesInfo = await ApiSingleton.filesApi.filesGetFilesInfo(+courseId!)
@@ -612,7 +201,7 @@ const Course: React.FC = () => {
     const unratedSolutionsCount = studentSolutions
         .flatMap(x => x.homeworks)
         .flatMap(x => x!.tasks)
-        .filter(t => t!.solution!.slice(-1)[0]?.state === 0)
+        .filter(t => t!.solution!.slice(-1)[0]?.state === 0) //last solution
         .length
 
     const [lecturerStatsState, setLecturerStatsState] = useState(false);
@@ -662,7 +251,7 @@ const Course: React.FC = () => {
                         </ListItemIcon>
                         <ListItemText>Поделиться</ListItemText>
                     </MenuItem>
-                    {isCourseMentor && 
+                    {isCourseMentor &&
                         <MenuItem onClick={() => {
                             setShowInviteDialog(true)
                         }}>
